// src/generated/models/agentDispatchStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * teaflask-serving
 * OpenAPI spec version: 1
 */

export type AgentDispatchStatus =
  (typeof AgentDispatchStatus)[keyof typeof AgentDispatchStatus];

export const AgentDispatchStatus = {
  dispatched: "dispatched",
  paused: "paused",
  succeeded: "succeeded",
  failed: "failed",
} as const;